uc-mcp-server/pyproject.toml
[project]
name = "uc-mcp-server"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0",
    "databricks-sdk>=0.30.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.scripts]
uc-mcp = "uc_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/uc_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pex>=2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
